+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: ageorgey <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/12/17 14:43:54 by ageorgey          #+#    #+#             */
-/*   Updated: 2018/11/25 15:58:19 by ageorgey         ###   ########.fr       */
+/*   Updated: 2018/12/03 18:41:57 by ageorgey         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -72,6 +72,6 @@ void				ft_putchar_fd(char c, int fd);
 void				ft_putstr_fd(char const *s, int fd);
 void				ft_putendl_fd(char const *s, int fd);
 void				ft_putnbr_fd(int n, int fd);
-int					ft_strnlen(char *str, char c);
+size_t				ft_strnlen(char *str, char c);
 
 #endif
